types: further improve Polymer interop types. (#4924)
This change adds better typing pattern for tf-paginated-view and
vz-histograms using the global ambient typing of 
`document.createElement`.
