+++ tensorboard/webapp/widgets/histogram/BUILD
@@ -12,6 +12,7 @@ tf_ng_module(
     ],
     deps = [
         ":types",
+        "//tensorboard/webapp:tb_polymer_interop_types",
         "@npm//@angular/core",
     ],
 )
@@ -26,6 +27,7 @@ tf_ts_library(
     deps = [
         ":histogram",
         ":types",
+        "//tensorboard/webapp:tb_polymer_interop_types",
         "//tensorboard/webapp/angular:expect_angular_core_testing",
         "//tensorboard/webapp/angular:expect_angular_platform_browser_animations",
         "@npm//@angular/core",
@@ -38,5 +40,8 @@ tf_ts_library(
     srcs = [
         "histogram_types.ts",
         "histogram_util.ts",
+    ],
+    deps = [
+        "//tensorboard/webapp:tb_polymer_interop_types",
     ],
 )
